style: split table.cs about tree node section
